Reorganize modules so methods related to datasets are together
