+++ cacimbao/datasets/__init__.py
@@ -1,0 +1,4 @@
+from .loaders import download_dataset, load_dataset
+from .metadata import list_datasets
+
+__all__ = ["download_dataset", "list_datasets", "load_dataset"]
